# services/user-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/sign-in": {
            "post": {
                "description": "Вход в аккаунт пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Вход в аккаунт",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignInResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-out": {
            "post": {
                "description": "Делает недействительным refresh токен и удаляет файл cookie refreshToken",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Выход из системы",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Missing or empty refreshToken cookie",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to logout",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Создание нового аккаунта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignUpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignInRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "ozon671games"
                },
                "password": {
                    "type": "string",
                    "example": "Da1dfshgn$"
                }
            }
        },
        "github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignInResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                }
            }
        },
        "github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignUpRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "Da1dfshgn$"
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 1,
                    "example": "ozon671games"
                }
            }
        },
        "github_com_SkySock_lode_libs_shared-dto_user_http_v1.SignUpResponse": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string",
                    "example": "01976451-00b3-7e32-9340-4f999c6c5edd"
                }
            }
        }
    }
}